# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "GoBup backend API for recording import, upload queues, publishing, danmaku processing, and system operations.",
        "title": "GoBup API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/progress/history/{historyId}": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Returns active part progress and an aggregate percentage for one history.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "progress"
                ],
                "summary": "Get history upload progress",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "History ID",
                        "name": "historyId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.HistoryProgressResponse"
                        }
                    }
                }
            }
        },
        "/progress/part/{partId}": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Returns current upload/transcoding progress for one part.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "progress"
                ],
                "summary": "Get part upload progress",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Part ID",
                        "name": "partId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controllers.HistoryProgressResponse": {
            "type": "object",
            "properties": {
                "activeCount": {
                    "type": "integer"
                },
                "historyId": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/upload.Progress"
                    }
                },
                "overallPercent": {
                    "type": "integer"
                }
            }
        },
...
        "upload.Progress": {
            "type": "object",
            "properties": {
                "chunkDone": {
                    "type": "integer"
                },
                "chunkTotal": {
                    "type": "integer"
                },
                "historyId": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "partId": {
                    "type": "integer"
                },
                "percent": {
                    "type": "integer"
                },
                "state": {
                    "$ref": "#/definitions/upload.ProgressState"
                },
                "stateMsg": {
                    "type": "string"
                },
                "updateAtMs": {
                    "type": "integer"
                }
            }
        },
        "upload.ProgressState": {
            "type": "string",
            "enum": [
                "TRANSCODING",
                "PROCESSING",
                "UPLOADING",
                "RETRY_WAIT",
                "SUCCESS",
                "FAILED"
            ],
            "x-enum-varnames": [
                "StateTranscoding",
                "StateProcessing",
                "StateUploading",
                "StateRetryWait",
                "StateSuccess",
                "StateFailed"
            ]
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}